refactor: renomme RéductionGénéraleMoisParMois
